pyproject.toml
[tool.poetry]
name = "derisk-solana"
version = "0.1.0"
description = "DeRisk - Risk monitoring and insurance against bad debt."
authors = ["Lukas Petrasek <lukas@carmine.finance>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
solana = ">=0.31.0"
sqlalchemy = "^2.0.25"
pandas = "^2.2.0"
flask = "^3.0.1"
flask-caching = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
streamlit = "^1.30.0"
plotly = "^5.18.0"
notebook = "6.5.6"  # This fixes the problem of jupyter showing just a blank page instead of the notebooks.
phoenix-trade = ">=0.1.9"
base58 = "^2.1.1"
asyncstdlib = "^3.12.0"
jsonrpcclient = "^4.0.3"
gfx-perp-sdk = "^0.0.15"
podite = "^0.1.2"
deepdiff = "^6.7.1"
borsh-construct = "^0.1.0"
anchorpy = "^0.19.1"
whirlpool-essentials = "^0.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pylint = "^3.0.3"
mypy = "^1.8.0"
types-requests = "^2.31.0.20240218"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
